Question 2/requirements.txt
psutil
cryptography
flask
flask-mysql
mysql-connector-python
